feat: Lotto 중복 데이터 유효성 기능 추가
